Message scrolling, dynamic chat, better objects.

+++ src/include/types.hpp
@@ -1,11 +1,11 @@
 #pragma once
 #include <string>
 
-typedef u_int16_t userid_t;
-typedef u_int16_t chatid_t;
-typedef u_int16_t code_t;
+typedef int userid_t;
+typedef int chatid_t;
+typedef int code_t;
 
-typedef u_int64_t objectid_t;
+typedef int objectid_t;
 
 typedef std::string login_t;
 typedef std::string password_t;
